Improve the Routing, splitting the Routers common methods to Core\BaseRouter
